src/conversation-api/requirements.txt
aiohttp==3.9.1
arxiv==2.0.0
azure-ai-contentsafety==1.0.0b1
azure-ai-formrecognizer==3.3.2
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-monitor-opentelemetry==1.1.0
azure-search-documents==11.4.0
fastapi==0.104.1
googlemaps==4.10.0
langchain==0.0.344
mmh3==4.0.1
numexpr==2.8.7
openai==1.3.6
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-system-metrics==0.41b0
opentelemetry-instrumentation-urllib3==0.41b0
pyjwt[crypto]==2.8.0
pyowm==3.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
qdrant-client==1.6.9
redis==5.0.1
sse-starlette==1.8.2
tenacity==8.2.3
tiktoken==0.5.1
uvicorn==0.24.0.post1
wikipedia==1.4.0
xmltodict==0.13.0
youtube-search==2.1.2
